Fix/AjusteProficiencia Release

+++ src/SME.SERAp.Prova.Infra/Dtos/Tai/ProficienciaAlunoProvaTaiDto.cs
@@ -6,9 +6,8 @@
         public long AlunoId { get; set; }
         public long AlunoRa { get; set; }
         public long? DisciplinaId { get; set; }
-        public long Proficiencia { get; set; }
+        public decimal Proficiencia { get; set; }
         public int Origem { get; set; }
         public int Tipo { get; set; }
-
     }
 }
